tj addon compat
